Move tests to their own test module
