+++ tests/project/test_project.py
@@ -11,7 +11,6 @@ from pyiron_base._tests import (
     PyironTestCase, TestWithProject, TestWithFilledProject, ToyJob
 )
 from pyiron_base.jobs.job.toolkit import BaseTools
-from pyiron_base.jobs.job.path import JobPath
 
 
 class TestProjectData(PyironTestCase):
@@ -128,43 +127,6 @@ class TestProjectOperations(TestWithFilledProject):
     def test_maintenance_get_repository_status(self):
         df = self.project.maintenance.get_repository_status()
         self.assertIn('pyiron_base', df.Module.values)
-
-    def test_load(self):
-        with self.subTest("Basic functionality"):
-            loaded = self.project.load("toy_1")
-            self.assertIsInstance(
-                loaded, ToyJob, msg="Expected to load the full object"
-            )
-
-            self.assertEqual(
-                loaded.name,
-                self.project.load.toy_1.name,
-                # TODO: we'd rather compare the jobs directly, but equality comparison
-                #       for jobs is not implemented
-                msg="Attribute access should function the same as explicit calls"
-            )
-
-            self.assertEqual(
-                len(self.project.job_table(recursive=False)),
-                len(self.project.load.__dir__()),
-                msg="Tab completion (`__dir__`) should see both jobs at this project "
-                    "level"
-            )
-
-        with self.subTest("Without converting to object"):
-            not_fully_loaded = self.project.load("toy_1", convert_to_object=False)
-            self.assertIsInstance(
-                not_fully_loaded,
-                JobPath,
-                msg="Expected to load only an HDF interface"
-            )
-
-            inspected = self.project.inspect("toy_1")
-            self.assertIsInstance(
-                inspected,
-                JobPath,
-                msg="Expected to load only an HDF interface"
-            )
 
 
 class TestToolRegistration(TestWithProject):
